refactoring & updates to qc
more


woops


style
