+++ visual_behavior/validation/utils.py
@@ -1,5 +1,32 @@
+import numpy as np
+import pandas as pd
+
 
 def assert_is_valid_dataframe(df, schema_instance):
     records = df.to_dict('records')
     errors = schema_instance.validate(records, many=True)
     assert (len(errors) == 0), errors
+
+
+def nanis_equal(v1, v2):
+    '''
+    checks equality, but deals with nulls (e.g., will return True for
+    np.nan==None)
+    '''
+    if pd.isnull(v1):
+        v1 = None
+    if pd.isnull(v2):
+        v2 = None
+    return v1 == v2
+
+
+def all_close(v, tolerance=0.01):
+    '''
+    adapts the numpy allclose method to work on single array
+    creates two arrays that are shifted versions of the input, pads with
+    initial and final values, compares. equivalent to iterating through v and
+    comparing each element to the next
+    '''
+    a = np.concatenate(([v[0]], v))
+    b = np.concatenate((v, [v[-1]]))
+    return np.allclose(a, b, rtol=tolerance)
